requirements.txt
fastapi
uvicorn
itsdangerous
pytest
allure-pytest
pydantic_settings
pre-commit
black
isort
flake8
httpx
